+++ src/modtool/info.py
@@ -1,0 +1,23 @@
+#
+# Copyright 2018 Free Software Foundation, Inc.
+#
+# This file is part of GNU Radio
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+#
+""" Returns information about a module """
+
+# Disallow running this script as a module
+if __name__ != '__main__':
+    exit(2)
+
+from sys import stderr
+from gnuradio.modtool.core import ModToolInfo, ModToolException
+
+try:
+    tool = ModToolInfo(python_readable=True)
+    tool.run()
+except ModToolException as e:
+    print(e, file=stderr)
+    exit(1)