+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the "gnuradio-integration" extension will be documented i
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Added
+- Added custom modtool wrapper scripts and `python-shell` dependency
+
+### Changed
+- Replaced `child_process` with `python-shell` for modtool commands
 
 ## [0.2.1] - 2023-12-07
 ### Added